[study] 자동차 경주 - PR review 내용 수정
