+++ src/main/java/domain/Car.java
@@ -3,6 +3,7 @@ package domain;
 public class Car {
     String name;
     int distance;
+    final int RANDOM_PIVOT_NUMBER = 4;
 
     public Car(String name) {
         this.name = name;
@@ -26,7 +27,7 @@ public class Car {
     }
 
     public boolean isMove(int randomNumber) {
-        if (randomNumber >= 4)
+        if (randomNumber >= RANDOM_PIVOT_NUMBER)
             return true;
 
         return false;
